improvement: removed FormToggle and JarbFormToggle
A toggle is not a good use case for in a form, as they should only be
used when they can immediately have effect.
Removed FormToggle and JarbFormToggle.
Moved Toggle to Core as it's not a form element anymore.

Also updated the story for the Toggle so the toggles actually work.

Closes #233
