+++ src/core/Toggle/Toggle.scss
@@ -67,11 +67,3 @@
     }
   }
 }
-
-.form-group {
-  .toggle-container {
-    position: relative;
-    top: -2px;
-    left: 1.25rem;
-  }
-}
