+++ src/index.test.ts
@@ -39,7 +39,6 @@ describe('index', () => {
         "FlashMessage": [Function],
         "FormButton": [Function],
         "FormError": [Function],
-        "FormToggle": [Function],
         "Icon": [Function],
         "IconPicker": [Function],
         "ImageUpload": [Function],
@@ -51,7 +50,6 @@ describe('index', () => {
         "JarbDateRangePicker": [Function],
         "JarbDateTimeInput": [Function],
         "JarbFileInput": [Function],
-        "JarbFormToggle": [Function],
         "JarbIconPicker": [Function],
         "JarbImageUpload": [Function],
         "JarbInput": [Function],
